requirements.txt
playwright
python-dotenv
discord.py
tk
